+++ recyclebin_unix.go
@@ -13,29 +13,16 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-//go:build windows
+//go:build !windows
 
 package main
 
 import (
 	"os"
-	"syscall"
 )
 
-func enableWindowsColors() {
-	// Enable ANSI color support on Windows 10+
-	kernel32 := syscall.NewLazyDLL("kernel32.dll")
-	setConsoleMode := kernel32.NewProc("SetConsoleMode")
-
-	var mode uint32
-	handle := syscall.Handle(os.Stdout.Fd())
-
-	// Get current console mode
-	syscall.GetConsoleMode(handle, &mode)
-
-	// Enable VIRTUAL_TERMINAL_PROCESSING (0x0004)
-	mode |= 0x0004
-
-	// Set new console mode
-	setConsoleMode.Call(uintptr(handle), uintptr(mode))
+// moveToRecycleBin on Unix systems just does a regular delete
+// (Trash support would require desktop environment specific implementations)
+func moveToRecycleBin(path string) error {
+	return os.Remove(path)
 }