+++ recyclebin_windows.go
@@ -1,0 +1,70 @@
+// Copyright (C) 2025 Berkay Yetgin
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+//go:build windows
+
+package main
+
+import (
+	"syscall"
+	"unsafe"
+)
+
+var (
+	shell32          = syscall.NewLazyDLL("shell32.dll")
+	shFileOperationW = shell32.NewProc("SHFileOperationW")
+)
+
+const (
+	FO_DELETE          = 0x0003
+	FOF_ALLOWUNDO      = 0x0040
+	FOF_NOCONFIRMATION = 0x0010
+	FOF_SILENT         = 0x0004
+)
+
+type SHFILEOPSTRUCTW struct {
+	Hwnd                  uintptr
+	WFunc                 uint32
+	PFrom                 *uint16
+	PTo                   *uint16
+	FFlags                uint16
+	FAnyOperationsAborted int32
+	HNameMappings         uintptr
+	LpszProgressTitle     *uint16
+}
+
+// moveToRecycleBin moves a file to the Windows Recycle Bin
+func moveToRecycleBin(path string) error {
+	// Convert path to UTF-16 with double null termination
+	from, err := syscall.UTF16PtrFromString(path)
+	if err != nil {
+		return err
+	}
+
+	// Prepare the structure
+	op := SHFILEOPSTRUCTW{
+		WFunc:  FO_DELETE,
+		PFrom:  from,
+		FFlags: FOF_ALLOWUNDO | FOF_NOCONFIRMATION | FOF_SILENT,
+	}
+
+	// Call SHFileOperationW
+	ret, _, _ := shFileOperationW.Call(uintptr(unsafe.Pointer(&op)))
+	if ret != 0 {
+		return syscall.Errno(ret)
+	}
+
+	return nil
+}